app/requirements.txt
ijson
psycopg2-binary
requests


